+++ terraform/core/core.tf
@@ -3,7 +3,7 @@ resource "helm_release" "core-application" {
   name       = "core-application"
   chart      = "${var.helm_repository}/core/application"
   namespace  = "core"
-  create_namespace = true
+  create_namespace = false
   timeout = var.helm_timeout
 
   set {
@@ -23,9 +23,13 @@ resource "helm_release" "core-application" {
     name  = "database.password"
     value = random_password.core_database_password.result
   }
+#   set {
+#     name = "service.password"
+#     value = random_password.service_password.result
+#   }
   set {
     name  = "s3.password"
-    value = random_password.s3_password.result
+    value = "minioadmin" #random_password.s3_password.result
   }
 
   set {
@@ -43,7 +47,7 @@ resource "helm_release" "core-frontend" {
   name             = "core-frontend"
   chart            = "${var.helm_repository}/core/frontend"
   namespace        = "core"
-  create_namespace = true
+  create_namespace = false
   timeout          = var.helm_timeout
 
   set {