+++ terraform/infra/dashboard/welcome-board.tf
@@ -3,8 +3,8 @@ resource "helm_release" "welcome-board" {
   name       = "welcome-board"
   chart      = "../../../helm/infra/welcome"
   version    = "1.1.2"
-  namespace  = "monitoring"
-  create_namespace = true
+  namespace  = "dashboard"
+  create_namespace = false
 
   set {
     name  = "ingress.hosts"