+++ terraform/infra/grafana/loki.tf
@@ -3,8 +3,8 @@ resource "helm_release" "loki" {
   name       = "loki"
   chart      = "loki-stack" #loki-stack is used to also get promtail
   version    = "v2.10.2"
-  namespace  = "monitoring"
-  create_namespace = true
+  namespace  = "grafana"
+  create_namespace = false
 
   set {
     name  = "grafana.enabled"