+++ terraform/infra/grafana/main.tf
@@ -8,4 +8,4 @@ provider "kubernetes" {
   config_path = "~/.kube/config"
 }
 
-#management.otlp.tracing.endpoint: "http://tempo-distributor.monitoring:4318/v1/traces"
+#management.otlp.tracing.endpoint: "http://tempo-distributor.grafana:4318/v1/traces"