+++ terraform/infra/ingress/apisix/apisix.tf
@@ -2,10 +2,10 @@ resource "helm_release" "apisix" {
   name       = "apisix"
   repository = "https://apache.github.io/apisix-helm-chart"
   chart      = "apisix"
-  version    = "2.7.0"
+  version    = "2.8.0"
   namespace  = "ingress-apisix"
   timeout    = "300"
-  create_namespace = true
+  create_namespace = false
 
   set {
     name  = "service.type"
@@ -45,7 +45,7 @@ resource "helm_release" "apisix" {
 
   set {
     name  = "apisix.pluginAttrs.opentelemetry.collector.address"
-    value = "tempo-distributor.monitoring:4318"
+    value = "tempo-distributor.grafana:4318"
   }
 
   set {