+++ terraform/infra/ingress/kong/kong.tf
@@ -2,10 +2,10 @@ resource "helm_release" "kong" {
   name             = "kong"
   repository       = "https://charts.konghq.com"
   chart            = "kong"
-  version          = "2.38.0"
+  version          = "2.39.0"
   namespace        = "kong"
   timeout          = "120"
-  create_namespace = true
+  create_namespace = false
 
   set {
    name  = "proxy.type"