+++ terraform/infra/istio/istio.tf
@@ -3,8 +3,8 @@ resource "helm_release" "istio-base" {
   repository = "https://istio-release.storage.googleapis.com/charts"
   chart      = "base"
   namespace  = "istio-system"
-  create_namespace = true
-  version    = "1.22.0"
+  create_namespace = false
+  version    = "1.22.1"
   wait       = true
 
   set {
@@ -18,8 +18,8 @@ resource "helm_release" "istio-istiod" {
   repository = "https://istio-release.storage.googleapis.com/charts"
   chart      = "istiod"
   namespace  = "istio-system"
-  create_namespace = true
-  version    = "1.22.0"
+  create_namespace = false
+  version    = "1.22.1"
   wait       = true
 
   depends_on = [helm_release.istio-base]
@@ -37,7 +37,7 @@ resource "helm_release" "istio-cni-kind" {
   repository = "https://istio-release.storage.googleapis.com/charts"
   chart      = "cni"
   namespace  = "istio-system"
-  create_namespace = true
+  create_namespace = false
   version    = "1.22.0"
   wait       = true
 
@@ -56,7 +56,7 @@ resource "helm_release" "istio-cni-microk8s" {
   repository = "https://istio-release.storage.googleapis.com/charts"
   chart      = "cni"
   namespace  = "istio-system"
-  create_namespace = true
+  create_namespace = false
   version    = "1.22.0"
   wait       = true
 
@@ -81,16 +81,9 @@ resource "helm_release" "istio-ztunnel" {
   repository = "https://istio-release.storage.googleapis.com/charts"
   chart      = "ztunnel"
   namespace  = "istio-system"
-  create_namespace = true
+  create_namespace = false
   version    = "1.22.0"
   wait       = true
 
   depends_on = [helm_release.istio-base]
-}
-
-resource "terraform_data" "istio" {
-  provisioner "local-exec" {
-    when = create
-    command = "./label.sh"
-  }
 }