+++ terraform/invoice/invoice-process.tf
@@ -1,9 +1,10 @@
 resource "helm_release" "invoice-process-application" {
+  #depends_on = [helm_release.redis]
   repository = var.helm_repository
   name       = "invoice-process-application"
   chart      = "${var.helm_repository}/invoice-process/application"
   namespace  = "invoice"
-  create_namespace = true
+  create_namespace = false
   timeout = var.helm_timeout
 
   set {
@@ -23,6 +24,14 @@ resource "helm_release" "invoice-process-application" {
   set {
     name = "service.password"
     value = random_password.service_password.result
+  }
+  set {
+    name  = "s3.password"
+    value = "minioadmin" #random_password.s3_password.result
+  }
+  set {
+    name  = "redis.password"
+    value = random_password.redis_password.result
   }
 
   set {