+++ doc/release-notes.md
@@ -1,3 +1,16 @@
+# 1.3.2
+- cleanup of event folder
+- network policies added
+- basic auth to core service added
+- kiali connection to tempo tracing added, spring application name now contains the Namespace
+- upgrade to apisix chart 2.8.0
+- upgrade to kind kubernetes 1.3.0
+- upgrade to kubernetes dashboard 6.0.8 (7.x has a lot of breaking changes)
+- Namespaces are now created ahead of time during initial install
+- moved dashboard to dashboard namespace
+- moved everything grafana to grafana namespace
+- added redis cluster to invoice
+
 # 1.3.1
 - switchable kafka and rabbitmq broker added